+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Tag with how many elements are in queue
 - Check if video is already on queue
 - Snackbar for displaying feedback messages
+- Individual button to remove video from queue
 
 ### Improvements